requirements.txt
Flask~=2.0.1
Werkzeug~=2.0.1
SQLAlchemy~=1.4.32
python-dotenv~=0.19.0
gunicorn
flask_sqlalchemy
flask_login
flask_scss
PyMySQL~=1.0.2